nieuwe workflow v4

+++ app/src/main/java/com/example/phoneforfilm/data/Conversation.kt
@@ -5,8 +5,8 @@ import androidx.room.PrimaryKey
 
 @Entity(tableName = "conversations")
 data class Conversation(
-    @PrimaryKey(autoGenerate = true) val id: Long = 0,
-    val contactId: Long,
+    @PrimaryKey(autoGenerate = true) val id: Int = 0,
+    val contactId: Int,
     val lastMessage: String,
     val timestamp: Long
 )
